test(event-sourcing): share the trace-processing pipeline fixtures
The span-coalescing test had copy-pasted buildTraceDeps out of the sharding
test, which is the duplication the repo rules call out, and it pushed the
new file past the 300-source-line bound. Move the pipeline deps and the span
payload builders into a sibling support module and have both tests read from
it. No behaviour change.
